requirements.txt
# Required packages for the Personalized Study Assistant
# Updated to use Google Gemini 2.5 Pro

# Core web framework
streamlit>=1.28.0

# Google Gemini AI
google-genai>=1.29.0

# LangChain ecosystem (for document processing and vector stores)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0

# Vector database
chromadb>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=1.0.0

# Embeddings and ML
sentence-transformers>=5.0.0
tiktoken>=0.5.0

# Data processing
pandas>=2.0.0
numpy<2  # Required for compatibility

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Legacy packages (keeping for compatibility)
azure-ai-inference
azure-core